feat: Add decorator IgnoreProperty
IgnoreProperty will tell the Converters service shouldn't serialize the property
in the final rendered JSON object.
Swagger will not generate the documentation for this property.

When a model is given as return parameters, the converters will follow the JSON Schema
(except if the IgnoreProperty is used). All properties not annotated with
a decorator (JsonSchema decorators) will not appear in the in the rendered JSON object.

Closes: #238, #257