chore: remove aws-lambda dependency
The `aws-lambda` package is actually a standalone CLI (`lambda`)
unrelated to the `aws-lambda` import path used in lambdas.

The types for imports of `"aws-lambda"` do in fact come from the
`@types/aws-lambda` package but do not describe the `aws-lambda` npm
package (rather AWS makes it available at run time).

This just removes the dependency.